feat(ui): add AtomProvider for theme and project ID
This commit introduces a new AtomProvider component responsible
for managing and providing the application's theme and project
ID using Jotai atoms.  The AtomProvider uses the
`theme-mode` value from local storage to set the light or
dark theme and allows for dynamic switching. It also provides
access to the project ID.